+++ src/utils-inl.h
@@ -1,7 +1,7 @@
 namespace bpftrace {
 
 inline std::string GetProviderFromPath(std::string path) {
-  int i = path.rfind("/");
+  size_t i = path.rfind("/");
   return (i != std::string::npos) ? path.substr(i + 1) : path;
 }
